% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{cite}
\alias{cite}
\title{Cite}
\usage{
cite()
}
\value{
citaion entry
}
\description{
Addin to insert BibTex citation in Rmd documents
}
\examples{
# To run the addin place your cursor
# in a Rmd document with a defined
# bibliography file

}

